+++ src/pages/LibrariesPage.vue
@@ -1,11 +1,23 @@
 <template>
   <q-page class="column q-pa-md bg-grey-1">
-    <h4
-      class="q-ma-none q-mt-md q-mb-lg"
-      data-cy="page_libraries_title"
-    >
-      {{ $t('LibrariesPage.text.title') }}
-    </h4>
+    <div class="row justify-center items-center">
+      <h4
+        class="q-ma-none q-mt-md q-mb-lg q-mr-auto"
+        data-cy="page_libraries_title"
+      >
+        {{ $t('LibrariesPage.text.title') }}
+      </h4>
+      <q-btn
+        outline
+        no-caps
+        color="primary"
+        class="bg-white"
+        data-cy="libraries_button_add"
+        :label="$t('LibrariesPage.text.add')"
+        :icon="$t('LibrariesPage.icon.add')"
+        @click="$router.push('/add-library')"
+      />
+    </div>
     <libraries-table
       @show="goToLibrary"
       @remove="openRemoveLibraryDialog"